python@3: add arm64 compatibility patches
This set of patches includes the following upstream pull requests:

- PR #21114, "Support `arm64` in Mac/Tools/pythonw"
  https://github.com/python/cpython/pull/21114

- PR #21224, "allow python to build for macosx-11.0-arm64"
  https://github.com/python/cpython/pull/21224

- PR #21249, "ctypes fixes for arm64 Mac OS"
  https://github.com/python/cpython/pull/21249

Adding them here is warranted as `python@3.8` is widely used as a
dependency, and the patch is required to enable testing dependent
formulae.

Note that these have been successfully tested for `python@3.8`
but not for `python`.

The patch directives should be surrounded by an `if Hardware::CPU.arm?`
block.