python@3: add arm64 compatibility patches
This set of patches includes the following upstream pull requests,
in this order:

- PR 20171, "Fix _tkinter use"
  https://github.com/python/cpython/pull/20171
  (prerequisite for patch #21249 to apply)

- PR 21114, "Support `arm64` in Mac/Tools/pythonw"
  https://github.com/python/cpython/pull/21114

- PR 21224, "allow python to build for macosx-11.0-arm64"
  https://github.com/python/cpython/pull/21224

- PR 21249, "ctypes fixes for arm64 Mac OS"
  https://github.com/python/cpython/pull/21249

The patches for 20171 and 21249 have been minimally modified in order
to backport them to 3.8.3.

Note that these have been successfully tested for `python@3.8`
but not for `python@3.7`.

The patch directive should be surrounded by an `if Hardware::CPU.arm?`
block.