nutcracker: add patch to use Homebrew `libyaml`
This should help resolve a CI failure seen in
Homebrew/homebrew-core#78294.

While we're here, I should mention that it looks like `nutcracker`
vendors a bunch of other things, which might be why all the other
distros are still on an older version at the moment.

That isn't causing us problems at the moment, however, so let's focus on
fixing the things that are.