bdw-gc: add multithreading patch
This is a backport of ivmai/bdwgc@5668de71107022a316ee967162bc16c10754b9ce
so that it applies to the v8.0.6 tag.

The commit is applied with no changes, except for the changes to
`win32_threads.c`. The changes to this file are not relevant for our
purposes, and the original commit no longer applies.

This patch can be removed in v8.2.0.